First working version of symbols resolver; Corrected test for symbol resolving

+++ KeystoneNET/OptionType.cs
@@ -1,6 +1,6 @@
 ﻿namespace KeystoneNET
 {
-    public enum OptionType : int
+    public enum OptionType : short
     {
         KS_OPT_SYNTAX = 1,
         KS_OPT_SYM_RESOLVER,
